Move the method enums into types.py to break the import cycle
